+++ crispy_print/crispy_print/doctype/crispy_typst_block/crispy_typst_block.js
@@ -1,8 +1,12 @@
 // Copyright (c) 2026, Agathodaemon and contributors
 // For license information, please see license.txt
 
-// frappe.ui.form.on("Crispy Typst Block", {
-// 	refresh(frm) {
+frappe.ui.form.on("Crispy Typst Block", {
+	refresh(frm) {
+		if (frm.is_new()) return;
 
-// 	},
-// });
+		frm.add_custom_button(__("Open Builder"), () => {
+			frappe.set_route("ctb-builder", frm.doc.name);
+		});
+	},
+});